pyproject.toml
[project]
name = "anorha-control"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.22.1",
    "gradio>=6.5.1",
    "imagehash>=4.3.2",
    "mss>=10.1.0",
    "numpy>=2.4.2",
    "opencv-python>=4.13.0.90",
    "pillow>=12.1.0",
    "playwright>=1.58.0",
    "pyautogui>=0.9.54",
    "pynput>=1.8.1",
    "pyobjc-framework-cocoa>=12.1",
    "pyobjc-framework-quartz>=12.1",
    "requests>=2.32.5",
    "timm>=1.0.24",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
]
